Resolve "vue ModifierEvenement"

Closes #13

See merge request 2234758/GetTogether!21

+++ app/src/main/java/com/inviteme/model/entities/Evenement.kt
@@ -27,5 +27,5 @@ data class Evenement(
     val dateAjout: Timestamp,
     val dateModification: Timestamp,
     @ColumnInfo(name = "lieu_id")
-    val lieuId: Int?
+    val lieuId: Long?
 )
